Saturn and Uranus Molecules
Okay, so Saturn and Jupiter molecule scenes are actually the layer of
liquid hydrogen just outside the cores of those planets cause we don't
know what's actually in the core. The Uranus core has water, methane and
ammonia, so I just approximated a sphere for those molecules. But it's
super colorful and cool looking! :-)